+++ src/renderer/render/media/video.ts
@@ -1,18 +1,22 @@
-import { html } from 'typed-dom';
+import { html, define } from 'typed-dom';
 import { Collection } from 'spica/collection';
+import { ObjectFromEntries } from 'spica/alias';
 
 const extensions = [
   '.webm',
   '.ogv',
 ];
 
-export function video(url: URL, alt: string, cache?: Collection<string, HTMLElement>): HTMLVideoElement | undefined {
+export function video(target: HTMLImageElement, url: URL, cache?: Collection<string, HTMLElement>): HTMLVideoElement | undefined {
   if (!extensions.includes(url.pathname.split(/(?=\.)/).pop()!)) return;
-  if (cache?.has(url.href)) return cache.get(url.href)!.cloneNode(true) as HTMLVideoElement;
+  if (cache?.has(url.href)) return define(
+    cache.get(url.href)!.cloneNode(true) as HTMLVideoElement,
+    ObjectFromEntries([...target.attributes]
+      .map(attr => [attr.name, attr.value])));
   const el = html('video', {
-    class: 'media',
-    src: url.href,
-    alt,
+    src: target.getAttribute('data-src'),
+    ...ObjectFromEntries([...target.attributes]
+      .map(attr => [attr.name, attr.value])),
     muted: '',
     controls: '',
     style: 'max-width: 100%;',