+++ src/renderer/render/media/youtube.ts
@@ -1,7 +1,7 @@
-import { html } from 'typed-dom';
+import { html, define } from 'typed-dom';
 import { Collection } from 'spica/collection';
 
-export function youtube(url: URL, cache?: Collection<string, HTMLElement>): HTMLElement | undefined {
+export function youtube(target: HTMLImageElement, url: URL, cache?: Collection<string, HTMLElement>): HTMLElement | undefined {
   let id: string;
   switch (url.origin) {
     case 'https://www.youtube.com':
@@ -17,7 +17,10 @@ export function youtube(url: URL, cache?: Collection<string, HTMLElement>): HTML
   }
   if (url.pathname.split('/').pop()!.includes('.')) return;
   if (cache?.has(url.href)) return cache.get(url.href)!.cloneNode(true) as HTMLElement;
-  const el = html('div', { class: 'media' }, [
+  if (cache?.has(url.href)) return define(
+    cache.get(url.href)!.cloneNode(true) as HTMLElement,
+    { class: target.className });
+  const el = html('div', { class: target.className }, [
     html('div', { style: 'position: relative; padding-top: 56.25%;' }, [
       html('iframe', {
         src: `https://www.youtube.com/embed/${id}`,